Reverting $$ for inline text as it breaks in PyPI
The original way was the best; I couldn't find a good way how to display inline maths on PyPI. Single $ doesn't work and $$ puts everything on a new line. Suggestions I found propose to use Unicode characters but this won't make it the same as equations. Anyways, it's important if it displays correctly on GitHub, not PyPI. 

(Side note, maybe the documentation could move to Github Wiki?)