Upgrading readme and PDAW weights analysis (#59)
* Upgrading readme and PDAW weights analysis

README.md updated with possibility to use polarization of the field (prompted by colleagues from the group).

PDAW analysis of the weights was not very consistent. I've now modified it such that it makes more sense.

* Revise laser field polarization note in README

GitHub's markdown is unable to show vectors and absolute values normally; it turns them into columns somehow. This stupid way of splitting equations into parts is the only way I could force GitHub to show the correct equation. There are other solutions, but they are typically not compatible with other renderers.